Split out the S3Uploader as a separate class
