+++ bag_unpacker/src/main/scala/uk/ac/wellcome/platform/archive/bagunpacker/services/Unpacker.scala
@@ -5,25 +5,19 @@ import java.nio.file.Paths
 import java.time.Instant
 
 import com.amazonaws.services.s3.AmazonS3
-import com.amazonaws.services.s3.model.ObjectMetadata
-import com.amazonaws.services.s3.transfer.{TransferManagerBuilder, Upload}
 import org.apache.commons.compress.archivers.ArchiveEntry
+import uk.ac.wellcome.platform.archive.bagunpacker.S3Uploader
 import uk.ac.wellcome.platform.archive.bagunpacker.models.UnpackSummary
 import uk.ac.wellcome.platform.archive.bagunpacker.storage.Archive
 import uk.ac.wellcome.platform.archive.common.ConvertibleToInputStream._
-import uk.ac.wellcome.platform.archive.common.operation.{
-  OperationFailure,
-  OperationResult
-}
+import uk.ac.wellcome.platform.archive.common.operation.{OperationFailure, OperationResult}
 import uk.ac.wellcome.storage.ObjectLocation
 
 import scala.concurrent.{ExecutionContext, Future}
 
 class Unpacker(implicit s3Client: AmazonS3, ec: ExecutionContext) {
 
-  private val transferManager = TransferManagerBuilder.standard
-    .withS3Client(s3Client)
-    .build
+  private val s3Uploader = new S3Uploader()
 
   def unpack(
     srcLocation: ObjectLocation,
@@ -81,17 +75,11 @@ class Unpacker(implicit s3Client: AmazonS3, ec: ExecutionContext) {
       )
     }
 
-    val metadata = new ObjectMetadata()
-    metadata.setContentLength(archiveEntrySize)
-
-    val upload: Upload = transferManager.upload(
-      uploadLocation.namespace,
-      uploadLocation.key,
-      inputStream,
-      metadata
+    s3Uploader.putObject(
+      inputStream = inputStream,
+      streamLength = archiveEntrySize,
+      uploadLocation = uploadLocation
     )
-
-    upload.waitForUploadResult()
 
     archiveEntrySize
   }
